move MainWindow into ui package
