+++ App.xaml
@@ -2,12 +2,12 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:PrettyScatter"
-             StartupUri="MainWindow.xaml">
+             StartupUri="Ui/Main/MainWindow.xaml">
     <Application.Resources>
         <!-- GridSplitter -->
         <Style x:Key="GridSplitterVerticalGripStyle" TargetType="{x:Type GridSplitter}">
             <Setter Property="VerticalAlignment" Value="Stretch" />
-            <Setter Property="FocusVisualStyle" Value="{x:Null}" />
+            <Setter Property="HorizontalAlignment" Value="Stretch" />
             <Setter Property="Cursor" Value="SizeWE" />
             <Setter Property="FocusVisualStyle" Value="{x:Null}" />
             <Setter Property="Background"
@@ -21,16 +21,15 @@
                                 Width="{TemplateBinding Width}">
                             <Border Height="50" Width="4" VerticalAlignment="Center"
                                     HorizontalAlignment="Center" BorderThickness="1 0"
-                                    Background="White" BorderBrush="#A0A0A0">
-                            </Border>
+                                    Background="Transparent" BorderBrush="#A0A0A0" />
                         </Border>
                     </ControlTemplate>
                 </Setter.Value>
             </Setter>
         </Style>
 
-        <Style x:Key="VerticalGridSplitter" TargetType="{x:Type GridSplitter}" BasedOn="{StaticResource GridSplitterVerticalGripStyle}">
-            <Setter Property="HorizontalAlignment" Value="Stretch" />
+        <Style x:Key="VerticalGridSplitter" TargetType="{x:Type GridSplitter}"
+               BasedOn="{StaticResource GridSplitterVerticalGripStyle}">
             <Setter Property="Width" Value="5" />
         </Style>
     </Application.Resources>
