+++ Ui/Main/MainWindow.xaml.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Diagnostics;
 using System.Drawing;
 using System.Linq;
 using System.Windows;
@@ -12,7 +11,7 @@ using PrettyScatter.Utils.Ext;
 using ScottPlot;
 using ScottPlot.Plottable;
 
-namespace PrettyScatter
+namespace PrettyScatter.Ui.Main
 {
     public partial class MainWindow : Window
     {
